🚀 feat (component): add main and update mongoose repository

+++ packages/repository/lib/mongoose-repository.ts
@@ -1,7 +1,11 @@
 import { Connection, FilterQuery, Model, Schema } from 'mongoose';
 import { Repository } from './repository';
+import { ObjectId } from '@pcp/object-id';
 
-export class MongooseRepository<TEntity> implements Repository<TEntity> {
+export class MongooseRepository<
+  TEntity extends { id: ObjectId } = { id: ObjectId }
+> implements Repository<TEntity>
+{
   private readonly _model: Model<TEntity>;
 
   constructor(connection: Connection, name: string, schema: Schema) {
@@ -18,23 +22,38 @@ export class MongooseRepository<TEntity> implements Repository<TEntity> {
 
   public async update(
     filter: FilterQuery<TEntity>,
-    data: Partial<Omit<TEntity, 'id'>>,
+    data: Partial<Omit<TEntity, 'id'>>
   ) {
-    await this.model.updateOne(filter, {
+    if (filter instanceof ObjectId) {
+      await this.model.updateOne(filter, {
+        ...data,
+        dateTimeLastUpdated: new Date(),
+      });
+
+      return;
+    }
+
+    await this.model.updateMany(filter, {
       ...data,
       dateTimeLastUpdated: new Date(),
     });
   }
 
   public async delete(filter: FilterQuery<TEntity>) {
-    await this.model.deleteOne(filter);
-  }
+    if (filter instanceof ObjectId) {
+      await this.model.deleteOne(filter);
 
-  public async findOne(filter: FilterQuery<TEntity>) {
-    return this.model.findOne(filter);
+      return;
+    }
+
+    await this.model.deleteMany(filter);
   }
 
   public async find(filter?: FilterQuery<TEntity>) {
+    if (filter instanceof ObjectId) {
+      return this.model.findOne(filter);
+    }
+
     return this.model.find(filter || {});
   }
 }
